Add tests for registry
