+++ tests/parallel/conftest.py
@@ -31,15 +31,6 @@ async def inference_pool_registry(settings: Settings) -> InferencePoolRegistry:
 
 
 @pytest.fixture
-async def sum_model(inference_pool: InferencePool, sum_model: MLModel) -> MLModel:
-    parallel_model = await inference_pool.load_model(sum_model)
-
-    yield parallel_model
-
-    await inference_pool.unload_model(sum_model)
-
-
-@pytest.fixture
 async def error_model(inference_pool: InferencePool, error_model: MLModel) -> MLModel:
     model = await inference_pool.load_model(error_model)
